separate location routes from app.js
