+++ app/models/post.js
@@ -7,4 +7,4 @@ var PostSchema   = new Schema({
     description: String
 });
 
-module.exports = mongoose.model('Post', PostSchema);
+module.exports   = mongoose.model('Post', PostSchema);
